Moved files around. Wrote structures centered around interfaces and an isolated navigation class.
Moved "samples" folder into "testing". Created a folder under "testing" called "structureOptions" to hold potential methods of organizing code. All code that used to be under "testing" has been moved to a folder within "structureOptions" called "inheritanceStructure".

Wrote two other potential code structures.
The first is based around a single drivetrain class that would inherit methods from various interfaces such as a webcam interface or an IMU interface. This didn't work and is left mostly unfinished.
The second is based on a hardware class that contains an IMU, a webcam, and a mecanum drivetrain which extends a normal drivetrain. There will be a seperate navigator class which is in charge of tracking the position of the robot based on the drivetrain, the IMU, and the webcam.